A few new options in the queue builder: putInBackground(), finalFlushTime() and maxItems() maxItems supports bounded queues

+++ curator-recipes/src/main/java/com/netflix/curator/framework/recipes/queue/DistributedQueue.java
@@ -79,8 +79,9 @@ public class DistributedQueue<T> implements QueueBase<T>
     private final ListenerContainer<QueuePutListener<T>> putListenerContainer = new ListenerContainer<QueuePutListener<T>>();
     private final AtomicInteger lastChildCount = new AtomicInteger(0);
     private final int maxItems;
+    private final int finalFlushMs;
+    private final boolean putInBackground;
 
-    private final AtomicInteger     maxItemsCurrentCount = new AtomicInteger(0);
     private final AtomicInteger     putCount = new AtomicInteger(0);
     private final CuratorListener   listener = new CuratorListener()
     {
@@ -92,7 +93,6 @@ public class DistributedQueue<T> implements QueueBase<T>
                 if ( event.getWatchedEvent().getType() == Watcher.Event.EventType.NodeChildrenChanged )
                 {
                     internalNotify();
-                    updateMaxItemsCount();
                 }
             }
         }
@@ -124,7 +124,9 @@ public class DistributedQueue<T> implements QueueBase<T>
             int minItemsBeforeRefresh,
             boolean refreshOnWatch,
             String lockPath,
-            int maxItems
+            int maxItems,
+            boolean putInBackground,
+            int finalFlushMs
         )
     {
         Preconditions.checkNotNull(client, "client cannot be null");
@@ -136,6 +138,7 @@ public class DistributedQueue<T> implements QueueBase<T>
 
         isProducerOnly = (consumer == null);
         this.lockPath = lockPath;
+        this.putInBackground = putInBackground;
         this.consumer = consumer;
         this.minItemsBeforeRefresh = minItemsBeforeRefresh;
         this.refreshOnWatch = refreshOnWatch;
@@ -144,7 +147,13 @@ public class DistributedQueue<T> implements QueueBase<T>
         this.queuePath = queuePath;
         this.executor = executor;
         this.maxItems = maxItems;
+        this.finalFlushMs = finalFlushMs;
         service = Executors.newSingleThreadExecutor(threadFactory);
+
+        if ( (maxItems != QueueBuilder.NOT_SET) && putInBackground )
+        {
+            log.warn("Bounded queues should set putInBackground(false) in the builder. Putting in the background will result in spotty maxItem consistency.");
+        }
     }
 
     /**
@@ -197,21 +206,31 @@ public class DistributedQueue<T> implements QueueBase<T>
                 }
             );
         }
-
-        updateMaxItemsCount();
     }
 
     @Override
     public void close() throws IOException
     {
-        if ( !state.compareAndSet(State.STARTED, State.STOPPED) )
+        if ( state.compareAndSet(State.STARTED, State.STOPPED) )
         {
-            throw new IllegalStateException();
-        }
+            if ( finalFlushMs > 0 )
+            {
+                try
+                {
+                    flushPuts(finalFlushMs, TimeUnit.MILLISECONDS);
+                }
+                catch ( InterruptedException e )
+                {
+                    Thread.currentThread().interrupt();
+                }
+            }
 
-        putListenerContainer.clear();
-        client.getCuratorListenable().removeListener(listener);
-        service.shutdownNow();
+            putListenerContainer.clear();
+            client.getCuratorListenable().removeListener(listener);
+            service.shutdownNow();
+
+            internalNotify();
+        }
     }
 
     /**
@@ -370,7 +389,49 @@ public class DistributedQueue<T> implements QueueBase<T>
 
         putCount.incrementAndGet();
         byte[]              bytes = ItemSerializer.serialize(multiItem, serializer);
-        BackgroundCallback  callback = new BackgroundCallback()
+        if ( putInBackground )
+        {
+            doPutInBackground(item, path, givenMultiItem, bytes);
+        }
+        else
+        {
+            doPutInForeground(item, path, givenMultiItem, bytes);
+        }
+        return true;
+    }
+
+    private void doPutInForeground(final T item, String path, final MultiItem<T> givenMultiItem, byte[] bytes) throws Exception
+    {
+        client.create().withMode(CreateMode.PERSISTENT_SEQUENTIAL).forPath(path, bytes);
+        synchronized(putCount)
+        {
+            putCount.decrementAndGet();
+            putCount.notifyAll();
+        }
+        putListenerContainer.forEach
+        (
+            new Function<QueuePutListener<T>, Void>()
+            {
+                @Override
+                public Void apply(QueuePutListener<T> listener)
+                {
+                    if ( item != null )
+                    {
+                        listener.putCompleted(item);
+                    }
+                    else
+                    {
+                        listener.putMultiCompleted(givenMultiItem);
+                    }
+                    return null;
+                }
+            }
+        );
+    }
+
+    private void doPutInBackground(final T item, String path, final MultiItem<T> givenMultiItem, byte[] bytes) throws Exception
+    {
+        BackgroundCallback callback = new BackgroundCallback()
         {
             @Override
             public void processResult(CuratorFramework client, CuratorEvent event) throws Exception
@@ -406,9 +467,7 @@ public class DistributedQueue<T> implements QueueBase<T>
             }
         };
         internalCreateNode(path, bytes, callback);
-        return true;
     }
-
 
     @VisibleForTesting
     void internalCreateNode(String path, byte[] bytes, BackgroundCallback callback) throws Exception
@@ -473,7 +532,7 @@ public class DistributedQueue<T> implements QueueBase<T>
         {
             synchronized(this)
             {
-                while ( lastChildCount.get() >= maxItems )
+                while ( (lastChildCount.get() >= maxItems) && (state.get() == State.STARTED) )
                 {
                     if ( hasMaxWait )
                     {
@@ -653,7 +712,10 @@ public class DistributedQueue<T> implements QueueBase<T>
             {
                 bytes = client.getData().storingStatIn(stat).forPath(itemPath);
             }
-            client.delete().withVersion(stat.getVersion()).forPath(itemPath);
+            if ( client.isStarted() )
+            {
+                client.delete().withVersion(stat.getVersion()).forPath(itemPath);
+            }
 
             if ( type == ProcessType.NORMAL )
             {
@@ -727,25 +789,5 @@ public class DistributedQueue<T> implements QueueBase<T>
         }
 
         return false;
-    }
-
-    private void updateMaxItemsCount() throws Exception
-    {
-        if ( maxItems != QueueBuilder.NOT_SET )
-        {
-            BackgroundCallback      callback = new BackgroundCallback()
-            {
-                @Override
-                public void processResult(CuratorFramework client, CuratorEvent event) throws Exception
-                {
-                    if ( event.getType() == CuratorEventType.EXISTS )
-                    {
-                        maxItemsCurrentCount.set(event.getStat().getNumChildren());
-                        internalNotify();
-                    }
-                }
-            };
-            client.checkExists().inBackground(callback).forPath(queuePath);
-        }
     }
 }
